Add CI workflow which runs step only if its previous attempt failed
The goal is to have each step not run during the rerun if its result
during previous run was a success.
This is to possibly split single `make test` in 'mini.nvim' intro a
separate steps for each module and not rerun module test if it was
already a success.